refactored the skill description
